+++ components/aws/sagemaker/tests/unit_tests/tests/commonv2/test_spec_input_parsers_v2.py
@@ -46,5 +46,4 @@ class SpecInputParsersTestCase(unittest.TestCase):
 
 
 if __name__ == "__main__":
-    # unittest.main()
-    par.yaml_or_json_list("{a: 1}")
+    unittest.main()
